--- a/.github/workflows/ubuntu_install.yml
+++ b/.github/workflows/ubuntu_install.yml
@@ -26,8 +26,7 @@ jobs:
     strategy:
       matrix:
         include:
-          - {shared: ON}
-          - {shared: OFF}
+          shared: [ON, OFF]
     steps:
       - uses: actions/checkout@692973e3d937129bcbf40652eb9f2f61becf3332 # v4.1.7
       - name: Setup Ninja