--- a/fuzz/build.sh
+++ b/fuzz/build.sh
@@ -6,7 +6,7 @@ mkdir build
 AMALGAMATE_OUTPUT_PATH=./build/singleheader python3 singleheader/amalgamate.py
 
 $CXX $CFLAGS $CXXFLAGS \
-     -std=c++17 \
+     -std=c++20 \
      -I build/singleheader \
      -c fuzz/parse.cc -o parse.o
 
@@ -14,7 +14,7 @@ $CXX $CFLAGS $CXXFLAGS $LIB_FUZZING_ENGINE parse.o \
      -o $OUT/parse
 
 $CXX $CFLAGS $CXXFLAGS \
-     -std=c++17 \
+     -std=c++20 \
      -I build/singleheader \
      -c fuzz/can_parse.cc -o can_parse.o
 
@@ -22,7 +22,7 @@ $CXX $CFLAGS $CXXFLAGS $LIB_FUZZING_ENGINE can_parse.o \
      -o $OUT/can_parse
 
 $CXX $CFLAGS $CXXFLAGS \
-     -std=c++17 \
+     -std=c++20 \
      -I build/singleheader \
      -c fuzz/idna.cc -o idna.o
 
@@ -30,7 +30,7 @@ $CXX $CFLAGS $CXXFLAGS $LIB_FUZZING_ENGINE idna.o \
      -o $OUT/idna
 
 $CXX $CFLAGS $CXXFLAGS \
-     -std=c++17 \
+     -std=c++20 \
      -I build/singleheader \
      -c fuzz/url_search_params.cc -o url_search_params.o
 
@@ -38,7 +38,7 @@ $CXX $CFLAGS $CXXFLAGS $LIB_FUZZING_ENGINE url_search_params.o \
      -o $OUT/url_search_params
 
 $CXX $CFLAGS $CXXFLAGS \
-     -std=c++17 \
+     -std=c++20 \
      -I build/singleheader \
      -c build/singleheader/ada.cpp -o ada.o