--- a/include/ada/helpers.h
+++ b/include/ada/helpers.h
@@ -88,8 +88,8 @@ ada_really_inline void remove_ascii_tab_or_newline(std::string& input) noexcept;
  * Return the substring from input going from index pos to the end.
  * This function cannot throw.
  */
-ada_really_inline std::string_view substring(std::string_view input,
-                                             size_t pos) noexcept;
+ada_really_inline constexpr std::string_view substring(std::string_view input,
+                                                       size_t pos) noexcept;
 
 /**
  * @private