--- a/include/ada/url-inl.h
+++ b/include/ada/url-inl.h
@@ -81,7 +81,7 @@ inline std::ostream &operator<<(std::ostream &out, const ada::url &u) {
     out.host_start = out.protocol_end;
     out.host_end = out.host_start;
 
-    if (!has_opaque_path && checkers::begins_with(path, "//")) {
+    if (!has_opaque_path && path.starts_with("//")) {
       // If url's host is null, url does not have an opaque path, url's path's
       // size is greater than 1, and url's path[0] is the empty string, then
       // append U+002F (/) followed by U+002E (.) to output.
@@ -196,7 +196,7 @@ inline void url::copy_scheme(const ada::url &u) {
     if (port.has_value()) {
       output += ":" + get_port();
     }
-  } else if (!has_opaque_path && checkers::begins_with(path, "//")) {
+  } else if (!has_opaque_path && path.starts_with("//")) {
     // If url's host is null, url does not have an opaque path, url's path's
     // size is greater than 1, and url's path[0] is the empty string, then
     // append U+002F (/) followed by U+002E (.) to output.