--- a/singleheader/README.md
+++ b/singleheader/README.md
@@ -11,13 +11,13 @@ This will create two new files (ada.h and ada.cpp).
 You can then compile the demo file as follows:
 
 ```
-c++ -std=c++17 -c demo.cpp
+c++ -std=c++20 -c demo.cpp
 ```
 
 It will produce a binary file (e.g., demo.o) which contains ada.cpp.
 
 ```
-c++ -std=c++17 -o demo demo.cpp
+c++ -std=c++20 -o demo demo.cpp
 ./demo
 ```
 
@@ -29,7 +29,7 @@ You may build and link using CMake (--target demo), because CMake can configure 
 You may also build a C executable.
 
 ```
-c++ -c ada.cpp -std=c++17
+c++ -c ada.cpp -std=c++20
 cc -c demo.c
 c++ demo.o ada.o -o cdemo
 ./cdemo