--- a/src/unicode.cpp
+++ b/src/unicode.cpp
@@ -156,7 +156,7 @@ ada_really_inline bool has_tabs_or_newline(
 // U+003F (?), U+0040 (@), U+005B ([), U+005C (\), U+005D (]), U+005E (^), or
 // U+007C (|).
 constexpr static std::array<uint8_t, 256> is_forbidden_host_code_point_table =
-    []() constexpr {
+    []() consteval {
       std::array<uint8_t, 256> result{};
       for (uint8_t c : {'\0', '\x09', '\x0a', '\x0d', ' ', '#', '/', ':', '<',
                         '>', '?', '@', '[', '\\', ']', '^', '|'}) {
@@ -171,7 +171,7 @@ ada_really_inline constexpr bool is_forbidden_host_code_point(
 }
 
 constexpr static std::array<uint8_t, 256> is_forbidden_domain_code_point_table =
-    []() constexpr {
+    []() consteval {
       std::array<uint8_t, 256> result{};
       for (uint8_t c : {'\0', '\x09', '\x0a', '\x0d', ' ', '#', '/', ':', '<',
                         '>', '?', '@', '[', '\\', ']', '^', '|', '%'}) {
@@ -251,7 +251,7 @@ contains_forbidden_domain_code_point_or_upper(const char* input,
 }
 
 // std::isalnum(c) || c == '+' || c == '-' || c == '.') is true for
-constexpr static std::array<bool, 256> is_alnum_plus_table = []() constexpr {
+constexpr static std::array<bool, 256> is_alnum_plus_table = []() consteval {
   std::array<bool, 256> result{};
   for (size_t c = 0; c < 256; c++) {
     result[c] = (c >= '0' && c <= '9') || (c >= 'a' && c <= 'z') ||