--- a/src/url_aggregator.cpp
+++ b/src/url_aggregator.cpp
@@ -315,8 +315,7 @@ bool url_aggregator::set_pathname(const std::string_view input) {
   }
   clear_pathname();
   parse_path(input);
-  if (checkers::begins_with(get_pathname(), "//") && !has_authority() &&
-      !has_dash_dot()) {
+  if (get_pathname().starts_with("//") && !has_authority() && !has_dash_dot()) {
     buffer.insert(components.pathname_start, "/.");
     components.pathname_start += 2;
   }