@lemire 2024-09-02
@anonrig I recommend you bump this version:

https://github.com/ada-url/ada/blob/59cc69351645ece086bb76b2fbdd2eb50614000e/CMakeLists.txt#L39