@anonrig 2024-09-02
hey @mikea, do you know by any chance how to fix the fuzzer (oss-fuzz) to compile with C++20? even after adding `-std=c++20`, it still uses `c++17`