@lemire 2024-09-02
@anonrig I suggest merging. This is a non-blocking technical issue.

*If* you introduced a bug (very unlikely) we can fix it later.