@anonrig 2024-09-02
@lemire before merging this, should we release a new version with @CarlosEduR's changes?