@anonrig 2024-09-03
I created a new release for v2.9.2. Once the github release is available, we can merge this.